First GC interface reshufflings to isolate G1 and CardTable barriers
